backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pydantic
python-multipart
python-dotenv
supabase
